[Refactor] Reorganised code structure (cont'd)
