+++ bertrend/demos/summarization/summarizer_app.py
@@ -7,9 +7,9 @@ import os
 import streamlit as st
 from loguru import logger
 
-from bertrend.summary.abstractive_summarizer import AbstractiveSummarizer
-from bertrend.summary.chatgpt_summarizer import GPTSummarizer
-from bertrend.summary.extractive_summarizer import (
+from bertrend.services.summary.abstractive_summarizer import AbstractiveSummarizer
+from bertrend.services.summary.chatgpt_summarizer import GPTSummarizer
+from bertrend.services.summary.extractive_summarizer import (
     EnhancedExtractiveSummarizer,
     ExtractiveSummarizer,
 )
