+++ bertrend/services/summarizer.py
