+++ bertrend/services/summary/__init__.py
@@ -2,5 +2,3 @@
 #  See AUTHORS.txt
 #  SPDX-License-Identifier: MPL-2.0
 #  This file is part of BERTrend.
-
-from bertrend.summary.chatgpt_summarizer import GPTSummarizer
